src/modules/invoice/usecase/find-invoice/find-invoice.ts
import InvoiceGateway from "../../gateway/invoice.gateway";
import { FindInvoiceInputDTO, FindInvoiceOutputDTO } from "./find-invoice.dto";

export default class FindInvoiceUseCase {
  private _invoiceRepository: InvoiceGateway

  constructor(invoiceRepository: InvoiceGateway) {
    this._invoiceRepository = invoiceRepository
  }

  async execute(input: FindInvoiceInputDTO): Promise<FindInvoiceOutputDTO> {
    const invoice = await this._invoiceRepository.find(input.id)

    return {
      id: invoice.id.id,
      name: invoice.name,
      document: invoice.document,
      address: {
        street: invoice.address.street,
        number: invoice.address.number,
        complement: invoice.address.complement,
        city: invoice.address.city,
        state: invoice.address.state,
        zipCode: invoice.address.zipCode
      },
      items: invoice.items.map((item) => ({
        id: item.id.id,
        name: item.name,
        price: item.price
      })),
      total: invoice.items.reduce((prev, curr) => curr.price + prev, 0),
      createdAt: invoice.createdAt
    }
  }
}